Moved s3 and file cabinets to their own modules
