+++ cabinets/__init__.py
@@ -1,3 +1,5 @@
-from .cabinet import Cabinets
+from cabinets.cabinet import Cabinets
+from cabinets.protocols.file import FileCabinet
+from cabinets.protocols.s3 import S3Cabinet
 
 __all__ = (Cabinets,)
